Major bug fix
Fix a bug that would cause Pokemon that are previously seen with
unknown expiration times to not be added to the sightings database when
they are later seen with a valid TTH, and would not add IVs to the
database when encountering a Pokemon that is not eligible for
notification. Thanks @ZeChrales.